Rewrite `Glossary` class to trait
