+++ src/ContaoManager/Plugin.php
@@ -23,6 +23,7 @@ use Contao\ManagerPlugin\Routing\RoutingPluginInterface;
 use Oveleon\ContaoGlossaryBundle\ContaoGlossaryBundle;
 use Symfony\Component\Config\Loader\LoaderResolverInterface;
 use Symfony\Component\HttpKernel\KernelInterface;
+use Symfony\Component\Routing\RouteCollection;
 
 /**
  * @internal
@@ -38,7 +39,7 @@ class Plugin implements BundlePluginInterface, RoutingPluginInterface
         ];
     }
 
-    public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel)
+    public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel): RouteCollection|null
     {
         return $resolver
             ->resolve('@ContaoGlossaryBundle/src/Controller')
